# Rewrite ColorMatrix