--- a/tests/ImageSharp.Benchmarks/Bulk/FromVector4_Rgb24.cs
+++ b/tests/ImageSharp.Benchmarks/Bulk/FromVector4_Rgb24.cs
@@ -4,7 +4,7 @@
 using BenchmarkDotNet.Attributes;
 using SixLabors.ImageSharp.PixelFormats;
 
-namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk;
+namespace SixLabors.ImageSharp.Benchmarks.Bulk;
 
 [Config(typeof(Config.ShortMultiFramework))]
 public class FromVector4_Rgb24 : FromVector4<Rgb24>