--- a/tests/ImageSharp.Benchmarks/Bulk/Pad3Shuffle4Channel.cs
+++ b/tests/ImageSharp.Benchmarks/Bulk/Pad3Shuffle4Channel.cs
@@ -3,7 +3,7 @@
 
 using BenchmarkDotNet.Attributes;
 
-namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk;
+namespace SixLabors.ImageSharp.Benchmarks.Bulk;
 
 [Config(typeof(Config.HwIntrinsics_SSE_AVX))]
 public class Pad3Shuffle4Channel