… tests/ImageSharp.Benchmarks/Bulk/PremultiplyVector4.cs
… tests/ImageSharp.Benchmarks/Bulk/Rgb24Bytes.cs
… tests/ImageSharp.Benchmarks/Bulk/ToRgba32Bytes.cs
… tests/ImageSharp.Benchmarks/Bulk/ToVector4.cs
… tests/ImageSharp.Benchmarks/Bulk/ToVector4_Bgra32.cs
… tests/ImageSharp.Benchmarks/Bulk/ToVector4_Rgb24.cs
… tests/ImageSharp.Benchmarks/Bulk/ToVector4_Rgba32.cs
… tests/ImageSharp.Benchmarks/Bulk/UnPremultiplyVector4.cs
… tests/ImageSharp.Benchmarks/Bulk/Vector4Factory.cs
… tests/ImageSharp.Benchmarks/General/GetSetPixel.cs
… tests/ImageSharp.Benchmarks/General/Vectorization/ColorNumerics.cs
… tests/ImageSharp.Benchmarks/ImageSharp.Benchmarks.csproj
… tests/ImageSharp.Benchmarks/Processing/BokehBlur.cs
… tests/ImageSharp.Tests/Primitives/ColorMatrixTests.cs