### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

Adopts the pattern introduced in .NET 8 by https://github.com/dotnet/runtime/pull/80091 for our `ColorMatrix` type.

A comfortable 2X speedup over main.

<!-- A description of the changes proposed in the pull-request -->

```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
11th Gen Intel Core i7-11370H 3.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.200
  [Host]     : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT
  DefaultJob : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT
```

Main

|         Method |     Mean |     Error |    StdDev | Ratio | RatioSD |
|--------------- |---------:|----------:|----------:|------:|--------:|
|      Transform | 8.413 us | 0.1002 us | 0.0888 us |  1.00 |    0.00 |
| Transform_Span | 8.406 us | 0.1590 us | 0.1488 us |  1.00 |    0.02 |

PR

|         Method |     Mean |     Error |    StdDev | Ratio | RatioSD |
|--------------- |---------:|----------:|----------:|------:|--------:|
|      Transform | 4.135 us | 0.0819 us | 0.1148 us |  1.00 |    0.00 |
| Transform_Span | 3.145 us | 0.0539 us | 0.0504 us |  0.76 |    0.02 |

<!-- Thanks for contributing to ImageSharp! -->