@JimBobSquarePants 2023-02-26
> > Ah... Bit of a problem here.
> > The compiled binary we ship will never contain the .NET 7 code. We target .NET 6 only and everything else will be stripped out.
> 
> Then you will also have problems with LossyUtils. There's also Conditional code.

Yeah. I did spot that after it was merged but it's only a couple of lines which serves as a reminder for the future, so I left it. This is a lot of extra code.