@JimBobSquarePants 2023-02-26
Is it just the `UnscopedRefAttribute` we're missing?

If so we can polyfill it. 

See https://github.com/Sergio0694/PolySharp/blob/615f6ccb3e59644e0e16752571c0db81ce780c5d/src/PolySharp.SourceGenerators/EmbeddedResources/System.Diagnostics.CodeAnalysis.UnscopedRefAttribute.cs

https://github.com/Sergio0694/PolySharp/blob/db53c4a2b3e931e0bf2489b0e405979421e44b64/tests/PolySharp.Tests/LanguageFeatures.cs#L127