@JimBobSquarePants 2023-02-26
If this passes, that's awesome! Only thing missing then is the transform.

https://github.com/dotnet/runtime/pull/80091/files#diff-4b222986f29a2b43c247d90ef0dba0c56b067f839a300583b7ed8ca0ac14d246R636-R647

Our equivalent is in ColorNumerics.