@JimBobSquarePants 2023-02-26
Th tests only fail on windows because we're not doing a comparison check on other platforms.